- Updated ogl-es branch to rev4182. - Added iOS project - Added mutli touch support (thanks to Redshift Software).
git-svn-id: svn://svn.code.sf.net/p/irrlicht/code/branches/ogl-es@4184 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ source/Irrlicht/CImageLoaderWAL.cpp
@@ -281,7 +281,7 @@ IImage* CImageLoaderLMP::loadImage(irr::io::IReadFile* file) const
 
 	// maybe palette file
 	u32 rawtexsize = header.width * header.height;
-	if ( rawtexsize + sizeof ( header ) != file->getSize() )
+	if ( rawtexsize + sizeof ( header ) != (u32)file->getSize() )
 		return 0;
 
 	u8 *rawtex = new u8 [ rawtexsize ];
